Move around Cosmos chain modules
