+++ crates/cosmos/cosmos-relayer/src/chain/impls/mod.rs
@@ -1,4 +1,9 @@
 pub mod connection_handshake_message;
 pub mod create_client_message;
+pub mod events;
+pub mod ext;
+pub mod fields;
+pub mod log;
 pub mod query_client_state;
 pub mod query_consensus_state;
+pub mod types;
