+++ crates/cosmos/cosmos-relayer/src/chain/impls/types.rs
@@ -22,8 +22,8 @@ use hermes_relayer_components::runtime::traits::runtime::ProvideRuntime;
 use hermes_relayer_components_extra::telemetry::traits::telemetry::HasTelemetry;
 use hermes_relayer_runtime::types::runtime::HermesRuntime;
 
+use crate::chain::components::CosmosChainComponents;
 use crate::contexts::chain::CosmosChain;
-use crate::impls::chain::component::CosmosChainComponents;
 use crate::types::telemetry::CosmosTelemetry;
 
 impl ProvideRuntime<CosmosChain> for CosmosChainComponents {
