--- crates/cosmos/cosmos-relayer/src/impls/chain/mod.rs
@@ -1,7 +1,0 @@
-pub mod component;
-pub mod components;
-pub mod events;
-pub mod ext;
-pub mod fields;
-pub mod log;
-pub mod types;
