+++ crates/cosmos/cosmos-relayer/src/impls/mod.rs
@@ -1,6 +1,5 @@
 pub mod birelay;
 pub mod build;
-pub mod chain;
 pub mod error;
 pub mod relay;
 pub mod subscription;
