+++ crates/cosmos/cosmos-relayer/src/lib.rs
@@ -8,6 +8,7 @@
 
 extern crate alloc;
 
+pub mod chain;
 pub mod contexts;
 pub mod impls;
 pub mod types;
