+++ crates/solomachine/solomachine-relayer/src/impls/chain/cosmos_components/connection_handshake_message.rs
@@ -2,7 +2,7 @@ use cgp_core::prelude::*;
 use cgp_core::{DelegateComponent, HasErrorType};
 use hermes_cosmos_client_components::traits::message::{CosmosMessage, ToCosmosMessage};
 use hermes_cosmos_client_components::types::messages::connection::open_try::CosmosConnectionOpenTryMessage;
-use hermes_cosmos_relayer::impls::chain::components::connection_handshake_message::DelegateCosmosConnectionHandshakeBuilder;
+use hermes_cosmos_relayer::chain::impls::connection_handshake_message::DelegateCosmosConnectionHandshakeBuilder;
 use hermes_cosmos_relayer::types::error::{BaseError, Error};
 use hermes_relayer_components::chain::traits::components::connection_handshake_message_builder::ConnectionHandshakeMessageBuilder;
 use hermes_relayer_components::chain::traits::types::connection::{
