+++ crates/solomachine/solomachine-relayer/src/impls/chain/cosmos_components/create_client_message.rs
@@ -2,7 +2,7 @@ use cgp_core::prelude::*;
 use cgp_core::{DelegateComponent, HasErrorType};
 use hermes_cosmos_client_components::traits::message::{CosmosMessage, ToCosmosMessage};
 use hermes_cosmos_client_components::types::messages::client::create::CosmosCreateClientMessage;
-use hermes_cosmos_relayer::impls::chain::components::create_client_message::DelegateCosmosCreateClientMessageBuilder;
+use hermes_cosmos_relayer::chain::impls::create_client_message::DelegateCosmosCreateClientMessageBuilder;
 use hermes_cosmos_relayer::types::error::Error;
 use hermes_relayer_components::chain::traits::components::create_client_message_builder::CreateClientMessageBuilder;
 use hermes_relayer_components::chain::traits::types::create_client::HasCreateClientPayload;
