+++ crates/solomachine/solomachine-relayer/src/impls/chain/cosmos_components/query_consensus_state.rs
@@ -1,7 +1,7 @@
 use cgp_core::prelude::*;
 use cgp_core::{DelegateComponent, HasErrorType};
 use hermes_cosmos_client_components::traits::rpc_client::HasRpcClient;
-use hermes_cosmos_relayer::impls::chain::components::query_consensus_state::DelegateCosmosConsensusStateQuerier;
+use hermes_cosmos_relayer::chain::impls::query_consensus_state::DelegateCosmosConsensusStateQuerier;
 use hermes_relayer_components::chain::traits::components::consensus_state_querier::ConsensusStateQuerier;
 use hermes_relayer_components::chain::traits::types::consensus_state::HasConsensusStateType;
 use hermes_relayer_components::chain::traits::types::height::HasHeightType;
